requirements.txt
boto3>=1.26.0
botocore>=1.29.0 
tqdm>=4.64.0
rich>=13.0.0
pyyaml>=6.0.0